+++ go.mod
@@ -1,0 +1,3 @@
+module example/user/tasty
+
+go 1.21.4
